Added Developer Mode and core temprature readings.
This is to help debug exactly why cores crash with a C7 during normal game play regardless of the game or core. The log will get cleared down every session but the core temperature will not. USB required to run.

Add a file called `RA_DEV_MODE` to the root of the USB to enable `Developer Mode` for RetroArch.